+++ src/Stack/Build.hs
@@ -14,7 +14,6 @@
 
 module Stack.Build
   (build
-  ,clean
   ,withLoadPackage
   ,mkBaseConfigOpts
   ,queryBuildInfo)
@@ -43,7 +42,6 @@ import qualified Data.Vector as V
 import qualified Data.Yaml as Yaml
 import           Network.HTTP.Client.Conduit (HasHttpManager)
 import           Path
-import           Path.IO
 import           Prelude hiding (FilePath, writeFile)
 import           Stack.Build.ConstructPlan
 import           Stack.Build.Execute
@@ -51,7 +49,6 @@ import           Stack.Build.Haddock
 import           Stack.Build.Installed
 import           Stack.Build.Source
 import           Stack.Build.Target
-import           Stack.Constants
 import           Stack.Fetch as Fetch
 import           Stack.GhcPkg
 import           Stack.Package
@@ -183,14 +180,6 @@ withLoadPackage menv inner = do
         , packageConfigCompilerVersion = envConfigCompilerVersion econfig
         , packageConfigPlatform = configPlatform (getConfig econfig)
         }
-
--- | Reset the build (remove Shake database and .gen files).
-clean :: (M env m) => m ()
-clean = do
-    econfig <- asks getEnvConfig
-    forM_
-        (Map.keys (envConfigPackages econfig))
-        (distDirFromDir >=> removeTreeIfExists)
 
 -- | Set the code page for this process as necessary. Only applies to Windows.
 -- See: https://github.com/commercialhaskell/stack/issues/738